pyproject.toml
[project]
name = "weekly-metrics-dashboard"
version = "0.1.0"
description = "Weekly SEO/GEO metrics dashboard for marketing team"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45.0",
    "pandas>=2.2.0",
    "plotly>=6.0.0",
    "openai>=1.30.0",
    "python-dotenv>=1.1.0",
    "google-api-python-client>=2.130.0",
    "google-analytics-data>=0.18.0",
    "google-auth>=2.29.0",
    "psycopg[binary]>=3.2.0",
    "psycopg_pool>=3.2.0",
    "alembic>=1.15.0",
    "sqlalchemy>=2.0.0",
]
